Added a ContextAware class. This takes a different approach than the ContextAware marker class that had been used up until a few days ago. The problem with the marker class is that its effects were apparent even if your class derives from ContextAware via some hierarchy of base classes.
The approach in this checkin uses a metaclass that proxies descriptors
defined in a class with appropriate ContextDescriptors. ContextAware
must be present explicitly and visibly in a class' base-classes to
trigger this context-aware behaviour.

If this appears to be too odd a behaviour for a base class, then
another approach is to make ContextAware not a base class, but instead
a form of "class advice", like implements() from zope.interface.
So, something like this:
  class Foo:
      thingsDefinedInThisClassShallBeContextAware()


git-svn-id: svn://svn.zope.org/repos/main@10471 62d5b8a3-27da-0310-9561-8e5933582275