Some minor stylistic cleanups.
Created an InstanceOnlyDescriptor to fix the problem of a
__Security_checker__ attribute of a class, intended for its instances
only, also acting for the class.

The InstanceOnlyDescriptor may be more generally applicable. It is written
and tested to act as a complete data descriptor, that is more than strictly
the minimum needed for this module. Perhaps it should be moved elsewhere?


git-svn-id: svn://svn.zope.org/repos/main@10480 62d5b8a3-27da-0310-9561-8e5933582275