Fixed collector entry 169
==============================================================
= Request - Entry #1 by Anonymous User on Jun 18, 2003 1:07 am

Go to a folder, check NO boxes and press the "Rename" button.

The view does not detect the lack of items checked (i.e. nothing to be
renamed) and goes in "rename mode".

This means only the "Apply" and "Canel" buttons are visible (until
such time as one or these buttons is pressed and normality returns)
==============================================================

Also:

- The error message was not being displayed.

- The error wasn't being displayed with a distinctive style.

- Removed unused contents_table macro.


git-svn-id: svn://svn.zope.org/repos/main@10588 62d5b8a3-27da-0310-9561-8e5933582275