A major change in the semantics of the ConfigurationRegistry class (and its interface).  When you deactivate an entry, the next entry is activated.  A None entry may occur (once) at any position.  The info() method now has an optional keep_dummy argument, which, when True, returns an info entry corresponding to the dummy None in the data.
This is done so that uninstalling a bundle has a more natural effect:
all the activations that were overridden by the bundle are reactivated
when the bundle is uninstalled.

It is possible that this breaks part of the UI; I've fixed the only
place in the UI that I'm aware of that uses this, the code for
activating services; but there may be other places.


git-svn-id: svn://svn.zope.org/repos/main@10592 62d5b8a3-27da-0310-9561-8e5933582275