We can't just use the checker that came in on the object we're context- wrapping. The wrapper might require other permissions.
I was bitten by this because I'm returning an object from publishTraverse
that needed to be context-decorated. On entry to ContextWrapper, the
object was security-proxied. On exit from ContextWrapper, the decorated
object was security-proxied using the checker it came in with.

The change I'm checking in is to detect this, and if necessary, use a
combined checker.


git-svn-id: svn://svn.zope.org/repos/main@10601 62d5b8a3-27da-0310-9561-8e5933582275