Backport of skinnable bugfix to 1.3 branch.  Original HEAD checkin message:
Fix __of__ method so that it does not remove previous acquisition wrappers. This behavior causes getPhysicalPath to return the wrong value when you acquire objects above the CMF site during traversal. The effect of this is invalid uid and path index entries in the portal_catalog when objects are indexed during such a traversal.


git-svn-id: svn://svn.zope.org/repos/main@35732 62d5b8a3-27da-0310-9561-8e5933582275