Hook up the local and global view services for interface queries.
getRegistrationsForInterface() propagates the query to the next service if it also supports IInterfaceBasedRegistry.

Add an adapter for the GlobalViewService to support IInterfaceBasedRegistry.  The adapter goes here, because the global service is in zope.component, which shouldn't depend on zope.app.

Convert the global views to registrations so that they can be passed back from getRegistrationsForInterface().


git-svn-id: svn://svn.zope.org/repos/main@10704 62d5b8a3-27da-0310-9561-8e5933582275