Made the 'Undo' tab considerably more useful. You can choose to see just your own transactions, or everyone's (default to just your own), you can see more than the last 20 transactions. zodb's undoInfo supported passing through filter specifications, but this wasn't hooked up properly.
There's an outstanding issue with the username that's in the zodb
transaction 'user' field - it's being concatenated together with a
'path' that's never set and defaults to '/ '. For now, all the
unpleasantness is isolated in undo_log.pt, along with a comment
with more details. Will try to fix this properly tomorrow.

A bunch of tests to check that the specification stuff is passed
through correctly. More tests to test_undo.py to provide better
test coverage are needed.


git-svn-id: svn://svn.zope.org/repos/main@10889 62d5b8a3-27da-0310-9561-8e5933582275