Implemented Object field types, and ObjectWidget to go with it. Object fields have Fields on them, and when included in a form view, their sub-fields fully participate in the form generate and editing. You'd use this kinda thing for generating an Address field.
The change required the removal of the apply_update method on EditView. It is
replaced by the calling of applyChanges on each widget (facilitated by the
applyWidgetsChanges function of zope.app.form.utility).

To make this sane, the ObjectWidget must be used via CustomWidget so we can
indicate which factory is to be called to generate the container for the
Object's fields. ObjectWidget and SequenceWidget also allow overriding of
the widgets used to render their sub-fields.

If this is all confusing, then see the new widgets.txt help file in the
zope/app/browser/form/ directory, and it may or may not help :)


git-svn-id: svn://svn.zope.org/repos/main@10931 62d5b8a3-27da-0310-9561-8e5933582275