okay, I actually hooked up the vocabulary directive and registry, so that we can do actual useful things with vocabularies.
I also had to fix a call that could not be run due to a context wrapper, so
I had to remove all proxies, which is probably not the right thing to do.
Fred, could you look at this?

Finally, it seems that I am the first one that seriously plays with
vocabularies, since it is impossible that vocabularies are usable as they
were before this checkin. Furthermore, many parts of the framework are in
a sorry state; for example the widget for the IVocabularyField requires
really an implementation of ITokenizedVocabularyField and ITokenizedTerm.
Why? This beats the heck out of me.

I really thought that after so many people told me they studied
vocabularies and that I should have a look, that they also wrote some semi-
serious code in it. But this cannot have been the case. I think studying
means "I read the docs." in this case.

Vocabularies are definitely not in a ready state for the first beta and
need to be added to the to do. (I will do that, as soon as I have the bug
tracker setup.)

Finally, I will check in some code tomorrow, which demonstrates how
vocabularies work, in one of the most common use cases (in my opinion).


git-svn-id: svn://svn.zope.org/repos/main@11007 62d5b8a3-27da-0310-9561-8e5933582275