Implemented relevant data for Processes (read workflows). You can specify a schema which contains the fields that should be available as relevant data. Once the schema is set, you are able to specify "get" and "set" permissions for each field. That;s it for the setup.
When you then create a Content Object that uses the process definition for
which you declared the schema, you will see that there is a form in the
'Workflows' tab of this content object that allows you to manipulate the
releavant data.

I also enhanced the XML Import/Export to support relevant data and their
permissions.

The easiest way to check all this functionality out is to uncomment the
include directives in browser/workflow/stateful/configure.zcml and
workflow/stateful/configure.zcml (both at the end of the file).

This completes my code and status review of the workflow package, which I
think is now ready for the beta.


git-svn-id: svn://svn.zope.org/repos/main@11098 62d5b8a3-27da-0310-9561-8e5933582275