Made menu, factory and form titles and labels MessageIDs instead of plain TextLines/Texts. That means that they'll be translated as soon as the translation data is in place.
The change affected most ZCML files and tests that make up ZCML code as
they had to specify an i18n_domain. This is 'zope' for all of Zope3.

The change also breaks many Zope3 products as they probably do not specify
an i18n_domain. A fix for them will follow immediately.


git-svn-id: svn://svn.zope.org/repos/main@11165 62d5b8a3-27da-0310-9561-8e5933582275