Made FSPythonScripts and FSPageTemplates unowned.
Until now, scripts and templates were owned by the owner of the
portal.  User permissions were intersected with the rights of the
owner of the portal.  The intent of the ownership machinery is to
prevent privileged users from executing trojan scripts, but in this
case the scripts can't theoretically be trojans.  They can only be
created by users who have filesystem privileges.  Therefore script
ownership in this case is not useful and causes unnecessary security
denials.


git-svn-id: svn://svn.zope.org/repos/main@35764 62d5b8a3-27da-0310-9561-8e5933582275