Added lazy unregistration to the object hub. It's currently possible to leave invalid object registrations when containers are deleted. These changes help to mitigate this problem until a more robust solution is in place. Details of the lazy unregistration are documented in test_objecthub.py.
These changes minimize the impact to the hub. More aggressive checks/unregistration of missing objects can be added if need be.

This change also includes the addition of a temporary hub method - unregisterMissingObjects - for unregistering invalid/missing objects. This functionality is exposed in the hub control form through a button the user can click to remove any missing objects.


git-svn-id: svn://svn.zope.org/repos/main@11233 62d5b8a3-27da-0310-9561-8e5933582275