Use __setitem__ rather than setObject to add items to containers.
No longer need source and target adapters for copy and move.

Updated to reflect that add and remove events extend move, not the
other way around.

Changed to reflect new naming conventions for copying.

Added additional setup code to the basic contents tests because
contents now needs to look at the clipboard to decide whether to
display a paste button.


git-svn-id: svn://svn.zope.org/repos/main@11626 62d5b8a3-27da-0310-9561-8e5933582275