Updated a test to prevent it from causing other tests to break.
A test was testing asserting interfaces for built-in types.  It tried
to clean up after itself, but it's cleanup caused a central interface
registry to be cleared that shouldn't have been.  This caused non-test
interface declarations to be lost.

Changed the test to be much more careful about returning the affected
registry to the state it was in before the test. Unfortunately, this
makes the test depend on some implementation details, although not
more so than before.


git-svn-id: svn://svn.zope.org/repos/main@11749 62d5b8a3-27da-0310-9561-8e5933582275