Fixed a bug in creating local sites.
A local site has a next attribute that referes to teh containing site.
It needs to make sire that the next attribute os not security wrapped.

Also, when a site is added, it needs to add itself as a subsite of
it's containing site. The method for doing this did not have a
secirity assertion.


git-svn-id: svn://svn.zope.org/repos/main@12001 62d5b8a3-27da-0310-9561-8e5933582275