Only one VirtualHostMonster is allowed per container.
When a VHM is activated, it adds the mapping 'VIRTUAL_URL_PARTS': (SERVER_URL, BASEPATH1, virtual_url_path) to the request's 'other' dictionary.  If BASEPATH1 is empty, it is omitted from the tuple.  The joined parts are also added under the key 'VIRTUAL_URL'.  Since the parts are evaluated before traversal continues, they will not reflect modifications to the path during traversal or by the addition of a default method such as 'index_html'.


git-svn-id: svn://svn.zope.org/repos/main@24244 62d5b8a3-27da-0310-9561-8e5933582275