Patch from kiko: Remember has_mac value across handle_read() calls.
The presence or absence of a mac bit is really part of the persistent
state of SizedMessageAsyncConnection.  If state is True when
handle_read() is entered, we get into the code to check the mac and
process the message without testing whether a mac is present.

I wish there was a test for this, but the code it's very difficult to
setup a test that calls handle_read() in just the right state.  A good
project, but not one I'm up for now.


git-svn-id: svn://svn.zope.org/repos/main@3141 62d5b8a3-27da-0310-9561-8e5933582275