Changes in the Extra and Annotations directories of a zsync checkout could not be committed because create_object() would use the temporary container provided by a serialization adapter as a location to perform adapter looks and other location-sensitive operations; a separate context needs to be used to provide location information when traversing objects in the Extra and Annotations directories since those ephemeral containers provide no locations in the Zope site.
The context which should be used is the innermost object 'o' from
which the root is reachable via __parent__ references.  This is the
only way these objects can perform adapter looks or resolve persistent
references to objects in the tree.  They will (still) not be able to
resolve "parent" references because the extra and annotation values
for object 'o' have no parent.


git-svn-id: svn://svn.zope.org/repos/main@12231 62d5b8a3-27da-0310-9561-8e5933582275