The workflow can now expand abstract groups into true groups.
Here is the idea behind this: there are many specific groups, but they
are all global; there are also many general groups, but they should all
be local.  DCWorkflow becomes the main component that maps between
global and local groups.

Also changed to storing group names instead of monikers, to avoid the
need to explain what a security moniker is. ;-)


git-svn-id: svn://svn.zope.org/repos/main@35950 62d5b8a3-27da-0310-9561-8e5933582275