Add a few useful bits.
1) You can now set an invariant in your interface, where an invariant is a
validation constraint on the interface as a whole (as opposed to the individual
field constraints in schema).  The invariant should simply be a callable that
accepts an object claiming to implement the interface, and raises Invalid or a
subclass thereof if the constraint is not met (the eventual intent is to fit
into the "error view" framework, but other systems can work for now).  It is
expected that objects validated by such invariants will work primarily or
exclusively with Attributes (schema fields) rather than method output; as such,
a pattern of validating data without an actual object may simply be to create a
dummy object, update the object's __dict__ with the data values, and have it
validated.

2) A convenience method is now attached to interfaces to aid in the use of
invariants: validateInvariants.  It validates object with all defined
invariants.  If the errors argument is None (the default), the method raises
the first Invalid error; if errors is a list, it appends all errors to the
list, and then raises Invalid with the errors as the first element of the
"args" tuple.

3) Any other tagged data for the interface can now be added by functions within
an interface--see how the invariant function works, and how the __init__ of the
InterfaceClass snarfs up the information.

An obvious goal for this infrastructure will be to integrate it into the schema/forms machinery.


git-svn-id: svn://svn.zope.org/repos/main@12294 62d5b8a3-27da-0310-9561-8e5933582275