Changed the way providedBy worked to fix a bug in handling old-style extension classes.
Note that the test is actually in Packages3/Interfaces, the facade
that emulated Z2 interfaces using Z3 interfaces. It was easier
to write the test there, because I could actually write it using
extension class. :) My attempts at simulating EC failed and are too
hard to understand, at least today.


git-svn-id: svn://svn.zope.org/repos/main@12297 62d5b8a3-27da-0310-9561-8e5933582275