Fixed a bug and added a test for UtilityService.getRegisteredMatching().
Reimplemented getRegisteredMatching() for the PresentationService. It is
not part of the API yet, but we should do that!

As far as I can tell, the zope.component package is in a bad shape due to
some refactorings. There are missing documentation strings everywhere and
there are many legacy methods hanging around.


git-svn-id: svn://svn.zope.org/repos/main@12316 62d5b8a3-27da-0310-9561-8e5933582275