Fixed test failures (not mine - I was just setting up stuff for a live demo tomorrow night and found them broken).
Note that the z.app.interpreter.python file had some "needed for
2.2" code - this did not work under 2.3. Indeed, as it was trying to
concatenate a code object and a newline, I can't see how it'd work
there, either. Since we don't run on less than 2.3, I can't see why
the code is needed.

If the package is meant to be usable under 2.2, feel free to revert
the change and replace it with code that still works on 2.3.


git-svn-id: svn://svn.zope.org/repos/main@12328 62d5b8a3-27da-0310-9561-8e5933582275