Rename transactionalUndo() to undo().
The old undo was not used by Zope and shouldn't have been used by any
other client.  The newly named undo() is the preferred version of
undo.

XXX DemoStorage didn't implement transactionalUndo, so now it doesn't
implementation undo() at all.  We need to replace it with the demo
storage from ZODB4.

There are a few changes related to ZODB4 removal in this checkin.


git-svn-id: svn://svn.zope.org/repos/main@3287 62d5b8a3-27da-0310-9561-8e5933582275