Update factory __call__ API.
Then, I also think that verifyObject does not behave right, when the interface
says __call__(self, *args, **kw). In this case __call__(self, arg1) should
validate true, since it fulfills the previous pattern; i.e. we need some
flexibility in verifyObject() for cases in which the signiture is not always
the same. This was particularly important for factories.



I deactivated verifyObject() for this reason and replaced it by the weaker
isImplementedBy().


git-svn-id: svn://svn.zope.org/repos/main@12810 62d5b8a3-27da-0310-9561-8e5933582275