Added an optional argument to the interface __call__ method. If the argument is provided, then the value provided will be returned if adaptation can't be done.  If a second argument isn't provided, then a TypeError is raised if an adapter can't be computed.
The main motivation for this change is to provide queryAdapter-like
semantics, while also keeping PEP-246 adapt-like semantics (call
__conform__ before checking whether the object already provides the
interface).

So now, you can use:

adapter = IFoo(ob, None)

much as you might have used queryAdapter before (but with PEP 246
adapt semantics).

I generally avoid methods like this, that change their behavior
based on the number of arguments provided, however, I decided to
do it in this case to be more consistent with the PEP 246 adapt
method, and because this is __call__, rather than a "get" method.
I used a trick to avoid using a global marker.  This trick uses nested
scopes to provide hidden "static" data for functions.


git-svn-id: svn://svn.zope.org/repos/main@12921 62d5b8a3-27da-0310-9561-8e5933582275