Added getFactoriesFor() function, since there is not compact equivalent to it anywhere.
Added deprecation warnings to getFactory and queryFactory, since they are
replaced by getUtility and queryUtility. Jim said, that there is not need to
deprecate them; if we decide that, then we should just remove the deprecation
warning. I changed all of the Zope code to use getUtility and queryUtility
though.


git-svn-id: svn://svn.zope.org/repos/main@12980 62d5b8a3-27da-0310-9561-8e5933582275