Remove _noncurrent instance variable.
It's clear in hindsight that the connection only loads non-current
revisions for objects that have been invalidated.  Thus, we never need
to track them to invalidate stale objects; they'll be invalidated
anyway.  And loadBefore() will never return a current revision,
because it would only be called if the current revision could not be
used.


git-svn-id: svn://svn.zope.org/repos/main@3396 62d5b8a3-27da-0310-9561-8e5933582275