Add support for showing the view's 'provided' interface. Fixed tests accordingly.
If the view has no name, it now says 'no name' instead of being blank.



Now that the proxyView function has a factory attribute, let's use it to get
factory information.



Also, use the SimpleView's first base as the factory class.



Simplification of _getFactoryData().



Only show the template path relative to the Zope 3 path and not the entire
real path.


git-svn-id: svn://svn.zope.org/repos/main@13367 62d5b8a3-27da-0310-9561-8e5933582275