Support handler in data structure.
Rename getFile() to getFileInfo() and return a dictionary containing line and
column information as well.



Restructured getInfo() so that it only returns the info value when it is a
string.



Implemented getHandler() to return the path and URL of the handler. If the
handler is a function, the link will point to the containing module.



Fix tests.


git-svn-id: svn://svn.zope.org/repos/main@13379 62d5b8a3-27da-0310-9561-8e5933582275