Add presumed speed optimization for pack().
Only open the file for unbuffered I/O after finishing the first phase
of pack.  The first phase gets its end-of-file position from the main
thread, so there's no possibility of reading a 'c' record.

Timings on Linux are inconclusive, but it seems like using buffered
I/O for the initial phase should be faster.


git-svn-id: svn://svn.zope.org/repos/main@3433 62d5b8a3-27da-0310-9561-8e5933582275