Add automatic synchronization for Connection objects.
This changes the behavior slightly for invalidations.  A Connection
will receive invalidations even if it didn't have any modified objects
in the transaction, e.g. db.undo() or a read-only transaction.  For
this to work, it assumes that the thread that called db.open() is the
thread that is going to use the connection.  If this isn't true, the
synch argument to db.open() can be used to disable the new feature.


git-svn-id: svn://svn.zope.org/repos/main@3468 62d5b8a3-27da-0310-9561-8e5933582275