Variant of a patch from Gintautas Miliauskas.
The old code used itertools.chain(), which didn't work as expected.
If the first of the two iterator grew after the chain had started
consuming the second iterator, the new element(s) in the first
iterator would never be consumed.

Fix this bug by eliminating the need for chaining.  Process all the
_added_during_commit objects at the end, each with its own
ObjectWriter instance.


git-svn-id: svn://svn.zope.org/repos/main@3500 62d5b8a3-27da-0310-9561-8e5933582275