Packages that need to provide C headers for use by other packages need to mark those headers in their SETUP.cfg file using the "header" setting.
When headers are so marked, they are shared with the rest of the
packages in a distribution by copying the headers into a shared
directory within the distribution that's provided to all the
extensions.  This allows C modules within the system to share headers;
distutils doesn't have a good general approach to dealing with C-level
dependencies between packages.


git-svn-id: svn://svn.zope.org/repos/main@25769 62d5b8a3-27da-0310-9561-8e5933582275