Convert the field collection behavior as described in http://mail.zope.org/pipermail/zope3-dev/2004-May/010797.html
The Sequence field is removed.  As I spoke with Stephan, it may be acceptable to add the Sequence field back in if it actually means something.  Sequence did mean it was iterable but container access API was not described.  It should be described.  IPythonSequence might describe this API.  A Sequence widget should require specification of the factory.

Set now specifies a sets.Set.

IChoiceSequence was removed.


git-svn-id: svn://svn.zope.org/repos/main@13834 62d5b8a3-27da-0310-9561-8e5933582275